add attrs class for base, activity and item. still wip.
