+++ reproschema/models/__init__.py
@@ -1,3 +1,4 @@
 from .protocol import Protocol
-from .activity import Activity
+# from .activity import Activity
 from .item import Item
+from .activity_new import Activity
