+++ reproschema/models/tests/data/activities/activity2_schema.jsonld
@@ -1,0 +1,34 @@
+{
+    "@context": "https://raw.githubusercontent.com/ReproNim/reproschema//contexts/generic",
+    "@type": "reproschema:Activity",
+    "@id": "newActivity1_schema.jsonld",
+    "prefLabel": "testing pref",
+    "altLabel": "test alt label",
+    "description": "trial",
+    "schemaVersion": "1.0.0-rc4",
+    "version": "0.0.1",
+    "preamble": {
+        "en": "this is test preamble"
+    },
+    "citation": "https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1495268/",
+    "image": {
+        "@type": "AudioObject",
+        "contentUrl": "http://example.com/sample-image.png"
+    },
+    "ui": {
+        "order": [
+            "./item_1"
+        ],
+        "addProperties": [
+            {
+                "variableName": "item_1",
+                "isAbout": "./item_1",
+                "isVis": true,
+                "requiredValue": true
+            }
+        ],
+        "allow": [
+            "reproschema:Skipped"
+        ]
+    }
+}
